+++ src/Symfony/Component/Console/Style/FormatterInterface.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Console\Helper\Formatter;
+namespace Symfony\Component\Console\Style;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
