+++ src/Symfony/Component/Console/Style/Standard/BlockFormatter.php
@@ -9,10 +9,11 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Console\Helper\Formatter;
+namespace Symfony\Component\Console\Style\Standard;
 
 use Symfony\Component\Console\Formatter\OutputFormatter;
 use Symfony\Component\Console\Helper\Helper;
+use Symfony\Component\Console\Style\FormatterInterface;
 
 /**
  * Formats a message as a block of text.
